Sync upstream

+++ commons/src/main/kotlin/com/simplemobiletools/commons/extensions/Point.kt
@@ -5,7 +5,7 @@ import android.graphics.Point
 fun Point.formatAsResolution() = "$x x $y ${getMPx()}"
 
 fun Point.getMPx(): String {
-    val px = x * y / 1000000.toFloat()
-    val rounded = Math.round(px * 10) / 10.toFloat()
+    val px = x * y / 1000000f
+    val rounded = Math.round(px * 10) / 10f
     return "(${rounded}MP)"
 }
